Install Term::ReadLine::Perl without tests to fix potential timeouts
During the installation of the Perl module Term::ReadLine::Perl, it pauses for operator
input during the tests. If this is unattended, this could potantially time out which
will crash the script.
If we install without tests, this berhaviour is bypassed. ReadKey seems well supported
and compatible with WSL, so we can tolerate this.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>