Consolidate Sudo requests
Originally there was a sudo code near the end of the script, which would often
require re-entering the users password due to time take by the preceeding sections.
There was no need for this part to be at the end, so consolidate it after the
sudo requests at the start of the script and the password will only be needed
at the start.

Signed-off-by: Grant Ramsay <seapagan@gmail.com>